+++ src/components/Editor/editor.module.css
@@ -1,0 +1,60 @@
+.editor {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+:local .h2 {
+  composes: H2 from global;
+  composes: mt-10 from global;
+  color: white;
+}
+.form {
+  composes: text from global;
+  composes: mt-3 from global;
+  composes: mb-10 from global;
+  display: flex;
+  flex-direction: column;
+  max-width: 540px;
+}
+
+:local .label {
+  composes: mt-3 from global;
+  color: white;
+}
+
+:local .labelTitle {
+  composes: ml-4 from global;
+}
+
+:local .buttonSubmit {
+  composes: mt-4 from global;
+  max-width: 180px;
+  align-self: flex-end;
+}
+:local .buttonText {
+  composes: pl-4 from global;
+  composes: pr-4 from global;
+  composes: text from global;
+  color: white;
+}
+
+textarea {
+  resize: none;
+  display: flex;
+}
+
+textarea::-webkit-scrollbar {
+    width: 10px;
+  }
+   
+  textarea::-webkit-scrollbar-thumb {
+    border-radius: 10px;
+    background-color:var(--accent-color);
+  }
+  
+  textarea::-webkit-scrollbar-track {
+    background-color:#5A5D66;
+    border-radius: 10px;
+    margin-top: 16px;
+    margin-bottom: 16px;
+  }
